JBDS-3875 Split gulpfile.js to multiple modules to simplify maintenance
Fix moves out functions to different files to be able to share them
between different guil related files.
